requirements.txt
Brotli==1.0.9
click==8.1.3
colorama==0.4.4
cycler==0.11.0
dash==2.4.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-gif-component==1.1.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.1.2
Flask-Compress==1.12
fonttools==4.33.3
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
plotly==5.8.0
postgres==4.0
psycopg2-binary==2.9.3
tenacity==8.0.1
threadpoolctl==3.1.0
Werkzeug==2.1.2
zipp==3.8.0
